Fix parse/validate split inconsistency
The abi module leaves report version checking for "validation" but still
checks the policy format unnecessarily in a parsing function.

This commit separates validation more concretely and adds it as a step
to the verify module's main function for checking reports.

+++ abi/amdsp.go
@@ -171,5 +171,5 @@ func (e SevFirmwareErr) Error() string {
 	if e.Status == GuestRequestInvalidLength {
 		return "too few extended guest request data pages"
 	}
-	return fmt.Sprintf("unexpected firmware status (see SEV API spec): %x", e.Status)
+	return fmt.Sprintf("unexpected firmware status (see SEV API spec): %x", uint64(e.Status))
 }
